#17 refactor: 업로드 페이지 컴포넌트로 내부 분리
