+++ RN-mobile/src/components/Upload/SubmitButton/style.tsx
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+import { BackGroundColor } from "constants/color";
+
+export const styles = StyleSheet.create({
+  container: {
+    paddingVertical: 14,
+    backgroundColor: BackGroundColor.GREEN,
+    borderRadius: 8,
+  },
+  btnText: {
+    color: "white",
+    fontSize: 20,
+    fontWeight: "600",
+    textAlign: "center",
+  },
+});
