+++ RN-mobile/src/screens/Upload/style.tsx
@@ -1,27 +1,21 @@
 import { StyleSheet } from "react-native";
-import { BackGroundColor, FontColor } from "constants/color";
+import { FontColor } from "constants/color";
 
 export const styles = StyleSheet.create({
   container: {
     paddingVertical: 15,
     paddingHorizontal: 15,
   },
-  inputBox: {
+});
+
+export const commonStyle = StyleSheet.create({
+  container: {
     gap: 8,
     marginBottom: 30,
-  },
-  priceLabelBox: {
-    flexDirection: "row",
-    alignItems: "center",
-    gap: 2,
   },
   label: {
     fontSize: 15,
     fontWeight: "600",
-  },
-  subLabel: {
-    fontSize: 14,
-    color: FontColor.gray,
   },
   textInput: {
     padding: 15,
@@ -29,41 +23,12 @@ export const styles = StyleSheet.create({
     backgroundColor: "white",
     borderRadius: 8,
   },
-  textArea: {
-    height: 140,
-    lineHeight: 23,
-  },
-  errorBox: {
-    flexDirection: "row",
-    alignItems: "center",
-    gap: 4,
-  },
-  errorMsg: {
-    fontSize: 14,
-    fontWeight: "500",
-    color: FontColor.WARNNING,
-    alignItems: "center",
-  },
   warningInput: {
     borderColor: FontColor.WARNNING,
     borderWidth: 1,
   },
-  submitBtn: {
-    paddingVertical: 14,
-    backgroundColor: BackGroundColor.GREEN,
-    borderRadius: 8,
-  },
-  btnText: {
-    color: "white",
-    fontSize: 20,
-    fontWeight: "600",
-    textAlign: "center",
-  },
-  priceBox: {
-    gap: 8,
-  },
-  priceLabel: {
-    fontSize: 14,
-    fontWeight: "500",
+  textArea: {
+    height: 140,
+    lineHeight: 23,
   },
 });
